moved selectedVideo & selectedCustomer to App level. Selection persists when you navigate to a different page
