+++ src/components/Customers.js
@@ -1,22 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './Customers.css';
-import DisplayCustomerDetail from './DisplayCustomerDetail';
 import CustomerList from './CustomerList';
 
-const Customers = (props) => {
-  const [selectedCustomer, setSelectedCustomer] = useState();
-
-  const setSelectedCustomerCallBack = (customer) => {
-    setSelectedCustomer(customer);
-  };
-
+const Customers = ({ setSelectedCustomerCallBack }) => {
   return (
     <div className="customers">
       <CustomerList setSelectedCustomerCallBack={ setSelectedCustomerCallBack } />
-      {
-        selectedCustomer &&
-        <DisplayCustomerDetail customer={ selectedCustomer } />
-      }
     </div>
   );
 };
