+++ src/components/Videos.js
@@ -1,22 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './Videos.css';
-import DisplayVideoDetail from './DisplayVideoDetail';
 import VideoList from './VideoList';
 
-const Videos = (props) => {
-  const [selectedVideo, setSelectedVideo] = useState();
-
-  const setSelectedVideoCallBack = (video) => {
-    setSelectedVideo(video);
-  };
-
+const Videos = ({ setSelectedVideoCallBack }) => {
   return (
     <div className="videos">
       <VideoList setSelectedVideoCallBack={ setSelectedVideoCallBack } />
-      {
-        selectedVideo &&
-        <DisplayVideoDetail video={ selectedVideo } />
-      }
     </div>
   );
 };
